requirements.txt
# Core Dependencies
notion-client>=2.2.1
pydantic>=2.0.0
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Code Quality and Validation  
ruff>=0.1.0
mypy>=1.5.0
types-requests>=2.31.0

# Utilities
rich>=13.0.0
python-slugify>=8.0.0
beautifulsoup4>=4.12.0
httpx>=0.24.0

# Optional: For enhanced features
python-dateutil>=2.8.0
pyyaml>=6.0.0